+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Added sound effects and particles when feeding a girl to give propper feedback
 - Added sounds when going out with a girl and giving gifts
 - Made it so that changing any block in a building causes it to do a rescan
+- Added Breast Milk
+- Made each girl say something when receiving gifts
 
 ## v0.6.1 
 - Fixed the issue where girls in villagers would have tiny breasts